==> part2/step3_generate_testdata.ipynb <==
import pandas as pd
import numpy as np
import random
import json
import os

from itertools import permutations
import matplotlib.pyplot as plt
from collections import defaultdict

abstracts_test_fname = "data/abstracts_test.csv"
abstracts_test_df = pd.read_csv(abstracts_test_fname,sep='\t')
abstracts_test_df.head()

entities_test_fname = "data/entities_test.csv"
entities_test_df = pd.read_csv(entities_test_fname,sep='\t')
entities_test_df.head()

test_id2title ={ abs_id: title for abs_id, title in abstracts_test_df[['abstract_id','title']].values }
test_id2abstr ={ abs_id: abs_text for abs_id, abs_text in abstracts_test_df[['abstract_id','abstract']].values }

print(len(abstracts_test_df))
print(len(entities_test_df))

# entity-type dictionary
entityid2type ={ entity_ids: type for entity_ids, type in entities_test_df[['entity_ids','type']].values }
mention2type = { mention: type for mention, type in entities_test_df[['mention','type']].values }
mention2entityid = { mention: entity_ids for mention, entity_ids in entities_test_df[['mention','entity_ids']].values }

entity_type_valid = [('GeneOrGeneProduct', 'GeneOrGeneProduct'),
                     ('GeneOrGeneProduct', 'DiseaseOrPhenotypicFeature'),
                     ('ChemicalEntity', 'DiseaseOrPhenotypicFeature'),
                     ('DiseaseOrPhenotypicFeature', 'GeneOrGeneProduct'),
                     ('SequenceVariant', 'DiseaseOrPhenotypicFeature'),
                     ('ChemicalEntity', 'GeneOrGeneProduct'),
                     ('DiseaseOrPhenotypicFeature', 'SequenceVariant'),
                     ('DiseaseOrPhenotypicFeature', 'ChemicalEntity'),
                     ('ChemicalEntity', 'ChemicalEntity'),
                     ('GeneOrGeneProduct', 'ChemicalEntity'),
                     ('SequenceVariant', 'ChemicalEntity'),
                     ('ChemicalEntity', 'SequenceVariant'),
                     ('SequenceVariant', 'SequenceVariant'),
                     ('SequenceVariant', 'GeneOrGeneProduct'),
                     ('GeneOrGeneProduct', 'SequenceVariant')]


### generete test dataset file
sentence2entitise = defaultdict(list)
sentence_infos = dict()
relation_candidates = []
for key, df in entities_test_df.groupby('abstract_id'):
    #if key == 1711760:
    text = test_id2title[key] + ' ' + test_id2abstr[key]
    sentences = text.split('. ')
    entity_infos = dict()
    mention2entityid = {}
    sen2ents = {}
    for i, row in df.iterrows():
        entity_infos[row['offset_start'], row['offset_finish']] = row['mention']
        mention2entityid[row['mention']] = row['entity_ids']
        #entity = row['mention']

    for idx in range(len(sentences)):
        if idx == 0:
            sent_left = 0
            sent_right = len(sentences[idx]+'. ')
        else:
            sent_left = sent_right #+ 1
            sent_right += len(sentences[idx]+'. ')
        for entity_pos in entity_infos.keys():
            if entity_pos[0] > sent_left and entity_pos[1] < sent_right:
                sentence2entitise[sentences[idx]].append(entity_infos[entity_pos])

    for s in sentences:
        entities =[]
        for m in mention2entityid:
            if m in s:
                entities.append(m)

        entities = list(set(entities))
        comb = permutations(entities, 2)
        for c in comb:
            if (mention2type[c[0]], mention2type[c[1]]) in entity_type_valid:
                entity_1_mention, entity_2_mention = c[0], c[1]
                entity_1_id, entity_2_id = mention2entityid[c[0]], mention2entityid[c[1]]
                relation_candidates.append([key, 
                                            s, 
                                            entity_1_mention, 
                                            entity_2_mention, 
                                            entity_1_id, 
                                            entity_2_id])
                # if in entity_type_valid
                #  else: 
                # if we do not remove the entity-entity which not in entity_type_valid
    



dataset_df =pd.DataFrame(relation_candidates, 
                         columns=['abstract_id','sentence','entity_1_mention','entity_2_mention','entity_1_id','entity_2_id'])
print('size of test data: ',len(dataset_df))

dataset_df.sample(5)

dataset_df.to_csv('./data/test_data2.csv',index=None)
