==> part2/step6_bert_predict-final.ipynb <==
"""Imports"""
import pandas as pd
import numpy as np
import random
import json
import csv
# importing the itertools library
import itertools
import torch
from transformers import BertTokenizer, BertForSequenceClassification

data_fname = './data/test_data2.csv'
test_df = pd.read_csv(data_fname)

#entities_fname = "../data/test/entities_test.csv"
#entities_df = pd.read_csv(entities_fname,sep='\t')

#entityid2type ={ entity_ids: type for entity_ids, type in entities_df[['entity_ids','type']].values }

"""Load models &  resources"""
model_dir = './outputs/biobert_models'
model_fname = model_dir + '/model_epoch_10/'
print('where is the model?: ',model_fname)

label2id = {
    'Association': 0,
    'Positive_Correlation': 1,
    'Negative_Correlation': 2,
    'Bind': 3,
    'Cotreatment': 4,
    'Comparison':5,
    'Drug_Interaction':6,
    'Conversion':7,
    'Negative':8
      }

id2label = {}
for key,value in label2id.items():
    id2label[value] = key


def classify_relation(row):
    text = row['sentence'] + str('[SEP]') + row['entity_1_mention'] + str('[SEP]') + row['entity_2_mention']
    input_ids = torch.tensor(tokenizer.encode(text, 
                                              add_special_tokens=True,
                                              max_length=128)).unsqueeze(0)  # Batch size 1
    labels = torch.tensor([1]).unsqueeze(0)  # Batch size 1

    input_ids= input_ids.to(device) 
    labels= labels.to(device)
    with torch.no_grad():
        outputs = model(input_ids, labels=labels)
    logits = outputs[1]
    result = np.argmax(logits.cpu().numpy(),axis=1)[0]
    
    label = id2label[result]

    return label

# Instantiate the Bert tokenizer
tokenizer = BertTokenizer.from_pretrained(model_fname, do_lower_case=False)
model = BertForSequenceClassification.from_pretrained(model_fname,num_labels=len(label2id))
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# send model to device
model.to(device);

test_df['type']= test_df.apply(lambda x: classify_relation(x), axis=1)

test_df.head()

#Predict novelity

"""Load models &  resources"""
model_dir = 'outputs/models'
model_fname = model_dir + '/bert_model_epoch6/'
print('where is the model?: ',model_fname)

label2id = {
    'Novel': 0,
    'No': 1,
      }

id2label = {}
for key,value in label2id.items():
    id2label[value] = key


def classify_novelity(row):
    text = row['sentence'] + str('[SEP]') + row['entity_1_mention'] + str('[SEP]') + row['entity_2_mention']
    input_ids = torch.tensor(tokenizer.encode(text, 
                                              add_special_tokens=True,
                                              max_length=128)).unsqueeze(0)  # Batch size 1
    labels = torch.tensor([1]).unsqueeze(0)  # Batch size 1

    input_ids= input_ids.to(device) 
    labels= labels.to(device)
    with torch.no_grad():
        outputs = model(input_ids, labels=labels)
    logits = outputs[1]
    result = np.argmax(logits.cpu().numpy(),axis=1)[0]
    
    label = id2label[result]

    return label

# Instantiate the Bert tokenizer
tokenizer = BertTokenizer.from_pretrained(model_fname, do_lower_case=False)
model = BertForSequenceClassification.from_pretrained(model_fname,num_labels=len(label2id))
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# send model to device
model.to(device);

test_df['novel']= test_df.apply(lambda x: classify_novelity(x), axis=1)

test_df['id'] = test_df.index

submission_df = test_df[['id','abstract_id','type','entity_1_id','entity_2_id','novel']]

submission_df.head()

len(submission_df)

submission_df = submission_df[submission_df['type']!='Negative']

len(submission_df)

submission_df.to_csv('./data/submission_26.csv',sep='\t',index=None)
